+++ internal/bundler/bundler_importstar_test.go
@@ -1088,8 +1088,8 @@ func TestExportSelfCommonJSMinified(t *testing.T) {
 		},
 		expected: map[string]string{
 			"/out.js": `// /entry.js
-var b = c((d, e) => {
-  e.exports = {foo: 123};
+var b = c((d, a) => {
+  a.exports = {foo: 123};
   console.log(b());
 });
 module.exports = b();